fix: Update lod parameter type in TextureShader for consistency
- Changed the type of the lod parameter from int to float in both function signatures and internal variable definitions to ensure consistency across shader functions.
- Adjusted the handling of lod to directly use the float type, improving clarity and type safety in texture sampling logic.

+++ src/foundation/materials/nodes/TextureShaderNode.ts
@@ -32,7 +32,7 @@ export class TextureShaderNode extends AbstractShaderNode {
     this.__inputs.push(new Socket('scale', CompositionType.Vec2, ComponentType.Float, Vector2.fromCopy2(1, 1)));
     this.__inputs.push(new Socket('offset', CompositionType.Vec2, ComponentType.Float, Vector2.fromCopy2(0, 0)));
     this.__inputs.push(new Socket('rotation', CompositionType.Scalar, ComponentType.Float, Scalar.fromCopyNumber(0)));
-    this.__inputs.push(new Socket('lod', CompositionType.Scalar, ComponentType.Int, Scalar.fromCopyNumber(-1)));
+    this.__inputs.push(new Socket('lod', CompositionType.Scalar, ComponentType.Float, Scalar.fromCopyNumber(-1)));
 
     this.__outputs.push(new Socket('rgba', CompositionType.Vec4, ComponentType.Float));
     this.__outputs.push(new Socket('rgb', CompositionType.Vec3, ComponentType.Float));
